Replace usage of ImmutabilityHelper with react-addons-update
This commit removes the last internal dependencies on `ImmutabilityHelper`. For now, it simply copies the methods that were used and imports directly from react-addons-update (in both touch-ripple.jsx and utils/styles.js). In the future, we should do a code search to see if all uses of `react-addons-update` are actually necessary.

Signed-off-by: Neil Gabbadon <neil.gabbadon@emikra.com>